requirements.txt
customtkinter
llama-cpp-python